+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This build is not intended for submission, but rather for the developers to keep
 - That means a fix for the player logic had to be made.
 - The dynamic rules are in development.
     - The house and hotel prices are now variables which can be changed.
+    - The events "Shop closure", "Housing crisis" and "Housing abundance" have been prepared.
 
 ## Known issues